% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multidimensional_spline.R
\name{generate_multidimensional_spline}
\alias{generate_multidimensional_spline}
\title{Generate a Multidimensional Quadratic Spline Signal with Additive Noise}
\usage{
generate_multidimensional_spline(ndims, Kseg, n_points, snr, seed = NULL)
}
\arguments{
\item{ndims}{Number of dimensions.}

\item{n_points}{Signal length.}

\item{snr}{Signal to noise ratio.}

\item{seed}{Random seed.}
}
\value{
A numeric vector representing the continuous spline signal with noise.
}
\description{
Generate a Multidimensional Quadratic Spline Signal with Additive Noise
}
\examples{
result <- list(p = c(0, 1, 2), v = c(1, 0.5, -0.5), a = c(0.1, -0.1, 0))
segments <- c(3, 2, 4)
signal <- generate_Qspline_signal(result, segments, n_points = 100, noise_sd = 0.05)
plot(signal, type = "l")
}
