% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidim_l1tf_constrained.R
\name{l1tf_constrained}
\alias{l1tf_constrained}
\title{Computes the changepoints for a quadratic fit of data with L1TF.}
\usage{
l1tf_constrained(singal, K)
}
\description{
Computes the changepoints for a quadratic fit of data with L1TF.
}
