% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineOP-package.R
\docType{package}
\name{splineOP-package}
\alias{splineOP}
\alias{splineOP-package}
\title{splineOP: computing the change-point location of a quadratic spline approximation}
\description{
This package computes a quadratic piecewise approximation (quadratic spline) using an approach inspired by dynamic programming. The algorithm relies on two key approximations. The state values at each time step are discretized, and a finite set of initial spline speeds is considered at the initial time. The resulting method has a quadratic time complexity with respect to both the number of time steps and the number of state values.
}
\author{
\strong{Maintainer}: Vincent Runge \email{vincent.runge@univ-evry.fr} [copyright holder]

Authors:
\itemize{
  \item Nicolas Cecchi \email{nicolas.cecchi1@ens-paris-saclay.fr}
}

}
\keyword{internal}
