% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_states}
\alias{square_states}
\title{Square Each Element of a States Matrix}
\usage{
square_states(data, beta, states)
}
\arguments{
\item{data}{Numeric vector of length N (currently unused).}

\item{beta}{Numeric penalty value (currently unused).}

\item{states}{Numeric matrix of size S x N.}
}
\value{
Numeric matrix of size S x N where each element is squared.
}
\description{
Given a matrix of states of size S x N, this function returns a new matrix
where each element is squared. Additional arguments `data` and `beta` are
included for future extensions but are not used in the squaring.
}
\examples{
library(Rcpp)
# Create example states matrix
states <- matrix(1:6, nrow = 2, ncol = 3)
data <- c(10, 20, 30)
beta <- 0.5
# Call the function
square_states(data, beta, states)

}
